refactor(wallet): extract Client::IdentityOps concern module (#578)
Moves the 6 public identity/certificate methods and 7 private helpers
from client.rb into a new Client::IdentityOps concern module.
